identity_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic[email]==2.5.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9